+++ example/lib/presentation/pull_to_refresh/character_list_view.dart
@@ -24,8 +24,10 @@ class _CharacterListViewState extends State<CharacterListView> {
     super.initState();
   }
 
-  void _fetchPage(int pageKey) {
-    RemoteApi.getCharacterList(pageKey, _pageSize).then((newItems) {
+  Future<void> _fetchPage(int pageKey) async {
+    try {
+      final newItems = await RemoteApi.getCharacterList(pageKey, _pageSize);
+
       final isLastPage = newItems.length < _pageSize;
       if (isLastPage) {
         _pagingController.appendLastPage(newItems);
@@ -33,9 +35,9 @@ class _CharacterListViewState extends State<CharacterListView> {
         final nextPageKey = pageKey + newItems.length;
         _pagingController.appendPage(newItems, nextPageKey);
       }
-    }).catchError((error) {
+    } catch(error) {
       _pagingController.error = error;
-    });
+    }
   }
 
   @override