+++ example/lib/presentation/search_snackbar/character_sliver_list.dart
@@ -44,9 +44,14 @@ class _CharacterSliverListState extends State<CharacterSliverList> {
     super.initState();
   }
 
-  void _fetchPage(pageKey) {
-    RemoteApi.getCharacterList(pageKey, _pageSize, searchTerm: _searchTerm)
-        .then((newItems) {
+  Future<void> _fetchPage(pageKey) async {
+    try {
+      final newItems = await RemoteApi.getCharacterList(
+        pageKey,
+        _pageSize,
+        searchTerm: _searchTerm,
+      );
+
       final isLastPage = newItems.length < _pageSize;
       if (isLastPage) {
         _pagingController.appendLastPage(newItems);
@@ -54,9 +59,9 @@ class _CharacterSliverListState extends State<CharacterSliverList> {
         final nextPageKey = pageKey + newItems.length;
         _pagingController.appendPage(newItems, nextPageKey);
       }
-    }).catchError((error) {
+    } catch(error) {
       _pagingController.error = error;
-    });
+    }
   }
 
   @override